debug

+++ test/zip.test.ts
@@ -139,14 +139,16 @@ describe("write vsix", () => {
       "test.txt": "test content",
     });
 
-    try {
-      accessSync(path, constants.W_OK);
-      existsSync(join(path, "pkg.vsix"));
-      console.error("BEFORE LSTAT SYNC");
-      lstatSync(join(path, "pkg.vsix"));
-      console.error("AFTER LSTAT SYNC");
-    } catch (err) {
-      console.error("TRY CATCH", err);
+    if (process.platform === "win32") {
+      try {
+        accessSync(path, constants.W_OK);
+        existsSync(join(path, "pkg.vsix"));
+        console.error("BEFORE LSTAT SYNC");
+        lstatSync(join(path, "pkg.vsix"));
+        console.error("AFTER LSTAT SYNC");
+      } catch (err) {
+        console.error("TRY CATCH", err);
+      }
     }
 
     await expect(writeVsix({
@@ -160,22 +162,24 @@ describe("write vsix", () => {
       ],
     })).rejects.toThrow();
 
-    try {
-      console.error("BEFORE LSTAT SYNC 2");
-      console.error(lstatSync(join(path, "pkg.vsix")));
-      await unlink(join(path, "pkg.vsix"));
-      console.error("AFTER UNLINK");
+    if (process.platform === "win32") {
+      try {
+        console.error("BEFORE LSTAT SYNC 2");
+        console.error(lstatSync(join(path, "pkg.vsix")));
+        await unlink(join(path, "pkg.vsix"));
+        console.error("AFTER UNLINK");
 
-      await writeFile(join(path, "pkg.vsix"), "hello world");
-    } catch (err) {
-      console.error("TRY CATCH 2", err);
-    }
+        await writeFile(join(path, "pkg.vsix"), "hello world");
+      } catch (err) {
+        console.error("TRY CATCH 2", err);
+      }
 
-    try {
-      console.error("BEFORE LSTAT SYNC 3");
-      console.error(lstatSync(join(path, "pkg.vsix")));
-    } catch (err) {
-      console.error("TRY CATCH 3", err);
+      try {
+        console.error("BEFORE LSTAT SYNC 3");
+        console.error(lstatSync(join(path, "pkg.vsix")));
+      } catch (err) {
+        console.error("TRY CATCH 3", err);
+      }
     }
 
     expect(existsSync(join(path, "pkg.vsix"))).toBe(false);
